+++ R/rescale.R
@@ -1,4 +1,4 @@
-# NORMALIZING
+# SCALING
 #' @include AllGenerics.R
 NULL
 
@@ -52,7 +52,7 @@ setMethod(
   definition = function(x, y, total = 1) {
     assert_length(y, length(x))
 
-    y <- (y * total) / sum(y)
+    y <- (y * total) / sum(abs(y))
     xy <- list(x = x, y = y)
     xy
   }
@@ -175,5 +175,35 @@ setMethod(
   definition = function(x, f, ...) {
     xy <- grDevices::xy.coords(x)
     methods::callGeneric(x = xy$x, y = xy$y, f = f, ...)
+  }
+)
+
+# SNV ==========================================================================
+#' @export
+#' @rdname rescale_snv
+#' @aliases rescale_snv,numeric,numeric-method
+setMethod(
+  f = "rescale_snv",
+  signature = signature(x = "numeric", y = "numeric"),
+  definition = function(x, y, ...) {
+    ## Validation
+    assert_length(y, length(x))
+
+    y <- (y - mean(y)) / stats::sd(y)
+
+    xy <- list(x = x, y = y)
+    xy
+  }
+)
+
+#' @export
+#' @rdname rescale_snv
+#' @aliases rescale_snv,ANY,missing-method
+setMethod(
+  f = "rescale_snv",
+  signature = signature(x = "ANY", y = "missing"),
+  definition = function(x, ...) {
+    xy <- grDevices::xy.coords(x)
+    methods::callGeneric(x = xy$x, y = xy$y, ...)
   }
 )